+++ example/lib/presentation/src/theme/src/ac_calendar_theme_data.dart
@@ -1,7 +1,9 @@
 import 'ac_day_theme_data.dart';
 import 'ac_month_picker_theme_data.dart';
 import 'ac_pages_calendar_header_theme_data.dart';
+import 'ac_time_input_theme_data.dart';
 import 'ac_titled_month_theme_data.dart';
+import 'ac_titled_time_theme_data.dart';
 import 'ac_week_theme_data.dart';
 import 'ac_wheel_picker_theme_data.dart';
 
@@ -12,6 +14,8 @@ abstract class ACCalendarThemeData {
   ACMonthPickerThemeData get monthPickerTheme;
   ACWheelPickerThemeData get wheelPickerTheme;
   ACTitledMonthThemeData get titledMonthTheme;
+  ACTimeInputThemeData get timeInputTheme;
+  ACTitledTimeThemeData get titledTimeTheme;
 
   ACCalendarThemeData copyWith();
 }
@@ -24,6 +28,8 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
     ACMonthPickerThemeData? monthPickerTheme,
     ACWheelPickerThemeData? wheelPickerTheme,
     ACTitledMonthThemeData? titledMonthTheme,
+    ACTimeInputThemeData? timeInputTheme,
+    ACTitledTimeThemeData? titledTimeTheme,
   }) => ACLightCalendarThemeData.raw(
     pagesCalendarHeaderTheme: pagesCalendarHeaderTheme ?? ACLightPagesCalendarHeaderThemeData(),
     dayTheme: dayTheme ?? ACLightDayThemeData(),
@@ -31,6 +37,8 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
     monthPickerTheme: monthPickerTheme ?? ACLightMonthPickerThemeData(),
     wheelPickerTheme: wheelPickerTheme ?? ACLightWheelPickerThemeData(),
     titledMonthTheme: titledMonthTheme ?? ACLightTitledMonthThemeData(),
+    timeInputTheme: timeInputTheme ?? ACLightTimeInputThemeData(),
+    titledTimeTheme: titledTimeTheme ?? ACLightTitledTimeThemeData(),
   );
 
   const ACLightCalendarThemeData.raw({
@@ -40,6 +48,8 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
     required this.monthPickerTheme,
     required this.wheelPickerTheme,
     required this.titledMonthTheme,
+    required this.timeInputTheme,
+    required this.titledTimeTheme,
   });
 
   @override
@@ -61,6 +71,12 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
   final ACTitledMonthThemeData titledMonthTheme;
 
   @override
+  final ACTimeInputThemeData timeInputTheme;
+
+  @override
+  final ACTitledTimeThemeData titledTimeTheme;
+
+  @override
   ACLightCalendarThemeData copyWith({
     ACPagesCalendarHeaderThemeData? pagesCalendarHeaderTheme,
     ACDayThemeData? dayTheme,
@@ -68,6 +84,8 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
     ACMonthPickerThemeData? monthPickerTheme,
     ACWheelPickerThemeData? wheelPickerTheme,
     ACTitledMonthThemeData? titledMonthTheme,
+    ACTimeInputThemeData? timeInputTheme,
+    ACTitledTimeThemeData? titledTimeTheme,
   }) => ACLightCalendarThemeData(
     pagesCalendarHeaderTheme: pagesCalendarHeaderTheme,
     dayTheme: dayTheme,
@@ -75,6 +93,8 @@ class ACLightCalendarThemeData implements ACCalendarThemeData {
     monthPickerTheme: monthPickerTheme,
     wheelPickerTheme: wheelPickerTheme,
     titledMonthTheme: titledMonthTheme,
+    timeInputTheme: timeInputTheme,
+    titledTimeTheme: titledTimeTheme,
   );
 }