+++ example/lib/presentation/src/widgets/widgets.dart
@@ -5,4 +5,5 @@ export 'src/ac_wheel_picker.dart';
 export 'src/calendar/calendar.dart';
 export 'src/month/month.dart';
 export 'src/scroll_view/scroll_view.dart';
+export 'src/time_input/time_input.dart';
 export 'src/vertical/vertical.dart';